fix some behavioural issues with ephemeral messages

+++ roid/components.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import pprint
 import re
 import asyncio
 import functools
@@ -231,31 +232,33 @@ class Component(OptionalAsyncCallable):
 
     async def __call__(self, interaction: Interaction) -> Any:
         try:
-            resp, parent = await self._invoke(interaction)
+            resp, parent, ephemeral = await self._invoke(interaction)
         except Exception as e:
             if self._on_error is None:
                 raise e from None
-            resp, parent = await self._invoke_error_handler(interaction, e)
+            resp, parent, ephemeral = await self._invoke_error_handler(interaction, e)
 
-        if resp.delete_parent and (parent is not None):
+        if (not ephemeral) and resp.delete_parent and (parent is not None):
             await self.app._http.delete_interaction_message(parent.token)
 
         return resp
 
-    async def _invoke(self, interaction: Interaction) -> Tuple[Response, Interaction]:
-        kwargs, parent = await self._get_kwargs(interaction)
+    async def _invoke(
+        self, interaction: Interaction
+    ) -> Tuple[Response, Interaction, bool]:
+        kwargs, parent, ephemeral = await self._get_kwargs(interaction)
 
         if self._callback_is_coro:
-            return await self._callback(**kwargs), parent
+            return await self._callback(**kwargs), parent, ephemeral
 
         partial = functools.partial(self._callback, **kwargs)
         loop = asyncio.get_running_loop()
-        return await loop.run_in_executor(None, partial), parent
+        return await loop.run_in_executor(None, partial), parent, ephemeral
 
     async def _get_kwargs(
         self,
         interaction: Interaction,
-    ) -> Tuple[dict, Optional[Interaction]]:
+    ) -> Tuple[dict, Optional[Interaction], bool]:
         _, *reference_id = interaction.data.custom_id.split(":", maxsplit=1)
 
         if len(reference_id) == 0:
@@ -283,4 +286,9 @@ class Component(OptionalAsyncCallable):
         if self._ctx.options is not None and interaction.data.values is not None:
             kwargs[self._target_options_parameter] = interaction.data.values
 
-        return kwargs, ctx and ctx.get("parent")
+        if ctx is None:
+            ephemeral = False
+        else:
+            ephemeral = ctx.get("ephemeral", False)
+
+        return kwargs, ctx and ctx.get("parent"), ephemeral
